+++ src/plugin.hpp
@@ -59,6 +59,7 @@ extern Model* modelSickoSampler;
 extern Model* modelSickoSampler2;
 extern Model* modelSimpleSeq4;
 extern Model* modelSlewer;
+extern Model* modelStepSeq;
 extern Model* modelSwitcher;
 extern Model* modelSwitcherSt;
 extern Model* modelSwitcher8;